backend/requirements.txt
autobahn==0.17.0
constantly==15.1.0
incremental==16.10.1
pyserial==3.2.1
six==1.10.0
Twisted==16.6.0
txaio==2.5.2
zope.interface==4.3.2
